+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Rename `download` command to `retrive testcases` (`download` remains as an alias)
 * Rename `restore` command to `retrive submissions`
 * Rename `list-langs` command to `retrive languages`
+* Replace `--only-scraped` with `--full`
 * Make `env` field more configurable.
 * Now it prints most of the information to stderr, not stdout.
 * Move `service`, `contest` and `language` fields of `snowchains.toml` to another file.